base32ct: extract `Alphabet` trait
Extracts a sealed trait which is used to hide the implementation details
for particular encodings, similar to the `Variant` trait in the
`base64ct` crate.
