+++ base32ct/src/lib.rs
@@ -46,216 +46,15 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
+mod alphabet;
 mod encoding;
 mod error;
 
 pub use crate::{
+    alphabet::{
+        base32::{Base32, Base32Unpadded},
+        base32upper::Base32Upper,
+    },
     encoding::Encoding,
     error::{Error, Result},
 };
-
-/// RFC4648 lower case Base32 encoding with `=` padding.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct Base32;
-
-impl Encoding for Base32 {
-    const PADDED: bool = true;
-
-    fn decode_5bits(byte: u8) -> i16 {
-        decode_5bits_lower(byte)
-    }
-
-    fn encode_5bits(src: u8) -> u8 {
-        encode_5bits_lower(src)
-    }
-}
-
-/// RFC4648 upper case Base32 encoding with `=` padding.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct Base32Upper;
-
-impl Encoding for Base32Upper {
-    const PADDED: bool = true;
-
-    fn decode_5bits(byte: u8) -> i16 {
-        decode_5bits_upper(byte)
-    }
-
-    fn encode_5bits(src: u8) -> u8 {
-        encode_5bits_upper(src)
-    }
-}
-
-/// RFC4648 lower case Base32 encoding *without* padding.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct Base32Unpadded;
-
-impl Encoding for Base32Unpadded {
-    const PADDED: bool = false;
-
-    fn decode_5bits(byte: u8) -> i16 {
-        decode_5bits_lower(byte)
-    }
-
-    fn encode_5bits(src: u8) -> u8 {
-        encode_5bits_lower(src)
-    }
-}
-
-/// Decode 5-bits of lower-case Base32.
-fn decode_5bits_lower(byte: u8) -> i16 {
-    let src = byte as i16;
-    let mut ret: i16 = -1;
-
-    // if (src > 96 && src < 123) ret += src - 97 + 1; // -64
-    ret += (((0x60 - src) & (src - 0x7b)) >> 8) & (src - 96);
-
-    // if (src > 0x31 && src < 0x38) ret += src - 24 + 1; // -23
-    ret += (((0x31 - src) & (src - 0x38)) >> 8) & (src - 23);
-
-    ret
-}
-
-/// Decode 5-bits of upper-case Base32.
-fn decode_5bits_upper(byte: u8) -> i16 {
-    let src = byte as i16;
-    let mut ret: i16 = -1;
-
-    // if (src > 64 && src < 91) ret += src - 65 + 1; // -64
-    ret += (((0x40 - src) & (src - 0x5b)) >> 8) & (src - 64);
-
-    // if ($src > 0x31 && $src < 0x38) $ret += $src - 24 + 1; // -23
-    ret += (((0x31 - src) & (src - 0x38)) >> 8) & (src - 23);
-
-    ret
-}
-
-/// Encode 5-bits of lower-case Base32.
-fn encode_5bits_lower(src: u8) -> u8 {
-    let mut diff: i16 = 0x61;
-
-    // if (src > 25) ret -= 72;
-    diff -= ((25i16 - src as i16) >> 8) & 73;
-
-    (src as i16 + diff) as u8
-}
-
-/// Encode 5-bits of upper-case Base32.
-fn encode_5bits_upper(src: u8) -> u8 {
-    let mut diff: i16 = 0x41;
-
-    // if ($src > 25) $ret -= 40;
-    diff -= ((25 - src as i16) >> 8) & 41;
-
-    (src as i16 + diff) as u8
-}
-
-#[cfg(all(test, feature = "alloc"))]
-mod tests {
-    use crate::{Base32, Base32Unpadded, Base32Upper, Encoding, Error};
-
-    #[derive(Debug)]
-    struct TestVector {
-        decoded: &'static [u8],
-        encoded: &'static str,
-    }
-
-    const LOWER_PADDED_VECTORS: &[TestVector] = &[
-        TestVector {
-            decoded: &[0],
-            encoded: "aa======",
-        },
-        TestVector {
-            decoded: &[1, 2, 3, 5, 9, 17, 33, 65, 129],
-            encoded: "aebagbijcequdai=",
-        },
-        TestVector {
-            decoded: &[32, 7],
-            encoded: "eadq====",
-        },
-    ];
-
-    const LOWER_UNPADDED_VECTORS: &[TestVector] = &[
-        TestVector {
-            decoded: &[0],
-            encoded: "aa",
-        },
-        TestVector {
-            decoded: &[1, 2, 3, 5, 9, 17, 33, 65, 129],
-            encoded: "aebagbijcequdai",
-        },
-        TestVector {
-            decoded: &[32, 7],
-            encoded: "eadq",
-        },
-    ];
-
-    const UPPER_PADDED_VECTORS: &[TestVector] = &[
-        TestVector {
-            decoded: &[0],
-            encoded: "AA======",
-        },
-        TestVector {
-            decoded: &[1, 2, 3, 5, 9, 17, 33, 65, 129],
-            encoded: "AEBAGBIJCEQUDAI=",
-        },
-        TestVector {
-            decoded: &[32, 7],
-            encoded: "EADQ====",
-        },
-    ];
-
-    #[test]
-    fn decode_valid_base32() {
-        for vector in LOWER_PADDED_VECTORS {
-            assert_eq!(&Base32::decode_vec(vector.encoded).unwrap(), vector.decoded);
-        }
-
-        for vector in LOWER_UNPADDED_VECTORS {
-            assert_eq!(
-                &Base32Unpadded::decode_vec(vector.encoded).unwrap(),
-                vector.decoded
-            );
-        }
-
-        for vector in UPPER_PADDED_VECTORS {
-            assert_eq!(
-                &Base32Upper::decode_vec(vector.encoded).unwrap(),
-                vector.decoded
-            );
-        }
-    }
-
-    #[test]
-    fn decode_padding_error() {
-        let truncated =
-            &LOWER_PADDED_VECTORS[0].encoded[..(&LOWER_PADDED_VECTORS[0].encoded.len() - 1)];
-        assert_eq!(Base32::decode_vec(truncated), Err(Error::InvalidEncoding));
-    }
-
-    #[test]
-    fn decode_range_error() {
-        assert_eq!(
-            Base32::decode_vec(core::str::from_utf8(&[0, 0, 0]).unwrap()),
-            Err(Error::InvalidEncoding)
-        );
-    }
-
-    #[test]
-    fn encode_base32() {
-        for vector in LOWER_PADDED_VECTORS {
-            assert_eq!(&Base32::encode_string(vector.decoded), vector.encoded);
-        }
-
-        for vector in LOWER_UNPADDED_VECTORS {
-            assert_eq!(
-                &Base32Unpadded::encode_string(vector.decoded),
-                vector.encoded
-            );
-        }
-
-        for vector in UPPER_PADDED_VECTORS {
-            assert_eq!(&Base32Upper::encode_string(vector.decoded), vector.encoded);
-        }
-    }
-}
